Refactor code into modules with frame caching
